# specs/001-crud-empleados/contracts/employees-openapi.yaml
openapi: 3.0.3
info:
  title: Employees CRUD API
  version: 1.0.0
  description: API para alta, consulta, actualización y eliminación física de empleados.
servers:
  - url: http://localhost:8080
security:
  - basicAuth: []
paths:
  /api/empleados:
    get:
      summary: Listar empleados (paginado)
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - in: query
          name: size
          required: false
          schema:
            type: integer
            minimum: 5
            maximum: 5
            default: 5
      responses:
        '200':
          description: Página de empleados
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoPageResponse'
        '400':
          description: Parámetros de paginación inválidos
    post:
      summary: Crear empleado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoCreateRequest'
      responses:
        '201':
          description: Empleado creado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Datos inválidos
  /api/empleados/{clave}:
    get:
      summary: Consultar empleado por clave
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Empleado encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '404':
          description: Empleado no encontrado
    put:
      summary: Actualizar empleado
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmpleadoUpdateRequest'
      responses:
        '200':
          description: Empleado actualizado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmpleadoResponse'
        '400':
          description: Datos inválidos
        '404':
          description: Empleado no encontrado
    delete:
      summary: Eliminar empleado (hard delete)
      parameters:
        - in: path
          name: clave
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Empleado eliminado
        '404':
          description: Empleado no encontrado
  /actuator/health:
    get:
      summary: Estado de salud
      security: []
      responses:
        '200':
          description: Servicio saludable
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    EmpleadoCreateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoUpdateRequest:
      type: object
      required: [nombre, direccion, telefono]
      properties:
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoResponse:
      type: object
      required: [clave, nombre, direccion, telefono]
      properties:
        clave:
          type: integer
          format: int64
        nombre:
          type: string
          maxLength: 100
        direccion:
          type: string
          maxLength: 100
        telefono:
          type: string
          maxLength: 100
    EmpleadoPageResponse:
      type: object
      required: [content, page, size, totalElements, totalPages]
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/EmpleadoResponse'
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 5
          maximum: 5
        totalElements:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0