# src/main/java/com/dsw01/practica02/dto/EmpleadoResponse.java
package com.dsw01.practica02.dto;

public class EmpleadoResponse {

    private Long clave;
    private String nombre;
    private String direccion;
    private String telefono;

    public EmpleadoResponse() {
    }

    public EmpleadoResponse(Long clave, String nombre, String direccion, String telefono) {
        this.clave = clave;
        this.nombre = nombre;
        this.direccion = direccion;
        this.telefono = telefono;
    }

    public Long getClave() {
        return clave;
    }

    public void setClave(Long clave) {
        this.clave = clave;
    }

    public String getNombre() {
        return nombre;
    }

    public void setNombre(String nombre) {
        this.nombre = nombre;
    }

    public String getDireccion() {
        return direccion;
    }

    public void setDireccion(String direccion) {
        this.direccion = direccion;
    }

    public String getTelefono() {
        return telefono;
    }

    public void setTelefono(String telefono) {
        this.telefono = telefono;
    }
}